### 2020-02-07 #15 @gabrieldansereau
[Re] Chaos in a Three-Species Food Chain
**Original article:** [Hastings, A., & Powell, T. (1991). Chaos in a Three-Species Food Chain. Ecology, 72(3), 896–903. https://doi.org/10.2307/1940591](https://doi.org/10.2307/1940591)

**Scientific domain:** Ecology
**Programming language:** Julia
**Suggested editor:** @karthik @delsuc

Tagging co-author @tpoisot
