### 2020-02-13 #15 @rougier
@opetchey Could you edit this submission in  Ecology (and become editor for ReScience C)? I can help you with the procedure.
