### 2020-02-28 #15 @rougier
@dmcglinn Could you edit this submission in Ecology (and become editor for ReScience C)? I can help you with the whole procedure for this article.
