### 2020-02-28 #15 @rougier
@jsta Could you review this submission? (and eventually become an editor for ReScience since we have a lot of submission in ecology)
