### 2020-02-28 #15 @trashbirdecology
I would be happy to review, but will likely not turn it around until *mid to late March*.

I am not familiar with Julia, so I will not be able to easily comment on the code, however, the code looks straightforward enough. Might want to ensure someone else can strictly review the code.
