### 2020-02-28 #15 @trashbirdecology
Could you please point me to the reviewer guidelines?
