### 2020-03-09 #15 @rougier
@ChrisRackauckas Could you review this submission? (reviewer guidelines at ttps://rescience.github.io/edit/)
