### 2020-03-10 #15 @rougier
@pboesu Could you review this submission? (reviewer guidelines at https://rescience.github.io/edit/)
