### 2020-03-13 #15 @rougier
I'm fine with this schedule. Also, would you be interested in becoming associate editor for ecology (after this review). We tend to have a fair number of submissions aroudn ecology and other editors are quite busy (or are themselves authors like @tpoisot).
