### 2020-08-08 #15 @tpoisot
Hi @trashbirdecology @karthik @rougier - it's all good -- we're living in the middle of a pandemic and trying to do whatever we can. No balls dropped by anyone as far as I'm concerned, these are just the times we live through. There'll be more reviewers at some point, meanwhile, take care.
