### 2020-08-10 #15 @rougier
@cboettig Could you review this submission for ReScience C: "[Re] Chaos in a Three-Species Food Chain"
