### 2020-08-10 #15 @cboettig
@rougier sure thing. things are nuts right now but I'll give a stab at it.
