### 2020-08-10 #15 @cboettig
I still need to run the code, but some quick comments on the text:

- overall this is excellent. The write-up is clear, detailed, and insightful.  The authors have not only reproduced the initial results, but have had to reverse-engineer some critical details in order to do so, including the estimating of the initial conditions used, as well as several results that were only alluded to but not actually described or shown in the original paper.
- I would be interested to know if the choice of the numerical integration algorithm (which the authors here document but I believe the original authors did not) would make much difference (for reasonable choices).  The authors comment on which method was automatically selected by Julia's routines, but it might be interesting to see some comment on whether this was important.

- minor edit: Discussion, first sentence: `s/sensible/sensitive`, and `s/sensibility/sensitivity` (though I really like the idea of extreme sensibility of a model).

I'll report back again when I get a chance to run the code (please bug me otherwise!)
