### 2020-08-10 #15 @cboettig
This is my first time running code in julia.  It was very pleasant :tophat:

The README.md is nicely precise so a novice can cut-and-paste to get things running.  I would suggest including some brief comment on the runtime & cpu/memory needs expected to run `main.jl` -- when running unfamiliar code I always have this moment of worry on how long it might take.  (Also, the README doesn't say how to actually _run_ `main.jl`, a quick google suggested I try `import("main.jl")` but maybe this was intended to run in non-interactive mode? I dunno.

Setup seemed to go smoothly (on a relatively vanilla Ubuntu 20.04 server).  Running `main.jl` generated a large number of warnings or errors?

First this repeated many times:

```
Error in your model specification or the true solution is unstable.
└ @ DiffEqBase ~/.julia/packages/DiffEqBase/gTsoC/src/integrator_interface.jl:333
┌ Warning: dt <= dtmin. Aborting. There is either an error in your model specification or the true solution is unstable.
└ @ DiffEqBase ~/.julia/packages/DiffEqBase/gTsoC/src/integrator_interface.jl:333
┌ Warning: dt <= dtmin. Aborting. There is either an error in your model specification or the true solution is unstable.
└ @ DiffEqBase ~/.julia/packages/DiffEqBase/gTsoC/src/integrator_interface.jl:333
┌ Warning: dt <= dtmin. Aborting. There is either an error in your model specification or the true solution is unstable.
```

Then what looks like a more mundane warning or error probably related to graphics display or something?

```

GKS: Open failed in routine OPEN_WS
GKS: GKS not in proper state. GKS must be either in the state WSOP or WSAC in routine ACTIVATE_WS
/home/cboettig/.julia/packages/GR/oiZD3/src/../deps/gr/bin/gksqt: error while loading shared libraries: libQt5Widgets.so.5: cannot open shared object file: No such file or directory
connect: Connection refused
GKS: can't connect to GKS socket application
Did you start 'gksqt'?
```

When I run `include("HPmodel.jl")` by itself, I get:

```
julia> # Load Hastings & Powell's model
       include("HPmodel.jl")
┌ Warning: Symbolic calculations could not initiate. Likely there's a function which is not differentiable by SymEngine.
└ @ ParameterizedFunctions ~/.julia/packages/ParameterizedFunctions/WcTrd/src/ode_def_opts.jl:244
(::var"##280"{var"#34#36",Nothing,var"#35#37",Nothing,Nothing,Nothing,Expr,Expr}) (generic function with 2 methods)
```

Both of these may be irrelivant as so far the figures it's made look the same.  (I'm still getting copies of the second message/error as it seems to still be running, but I can see from the git diff new versions of the figures that look much like the originals).  If these warnings (particulary the first one) are expected and negligble, some documentation of that reasoning would be good.  Otherwise if I've done something wrong, and ideas / suggestions?

Also, minor point, but the `png` figures look pretty grainy in the pdf.  Perhaps the size would be prohibitive, but maybe consider using a vector graphics engine (may need to downsample the number of points down to what can be seen by a human eye) or just use a higher resolution for the pngs?)  obviously a very minor issue though.
