### 2020-08-10 #15 @cboettig
Final remarks:

Overall the Julia code looks very clean and easy to read.  Given that I seem to still get the approximately the same desired figures out, (unless there's some intermediate cache I didn't notice? the above errors/warnings are annoying but minor issues. Unless the authors or reviewers decide otherwise, I'm happy to leave it to them on how best (if at all) to address the issues I've raised above before final acceptance, and don't think I need to re-review any changes.  I enjoyed this piece and am happy to recommend it for publication.
