### 2020-08-11 #15 @gabrieldansereau
Thank you for your review @cboettig. I am very happy you could get up and running in Julia so quickly.

I will address your comments with some changes as soon as possible. But briefly:
- You raise good points on usability (runtime & needs, how to run the script). I'll add more info to clarify
- Interesting point on the choice of integration algorithm -- I'll re-explore and report back on it.
- The errors and warnings are expected and negligible, as you suspected (I had the same). The first is a warning about the integration itself, and I don't think there's much to do about it. The second is a graphics display error I sometimes get too on Ubuntu, but the figures are produced correctly anyways. The last one occurs when precompiling the packages -- again I don't think there's much to do. I'll check again if I can do something, else I'll add some comments to document.
- Thanks for noting the sensitivity typo and the figures' grainy look. I'll fix them and proof-read again.
