### 2020-08-11 7b0e993
clarify how to run 'main.jl'
