### 2020-08-11 e38bce8
produce GIF separately from 'main.jl'
