### 2020-08-11 6152ff5
clarify runtime & memory needs
