+++ article/content.tex
@@ -248,11 +248,11 @@ was similar to the one in \citeauthor{hastings1991}'s paper.
 \section{Discussion}
 
 We were able to replicate the chaotic behaviour displayed by \citeauthor{hastings1991}'s model.
-The resulting behaviour is indeed very sensible to the initial conditions, showing
+The resulting behaviour is indeed very sensitive to the initial conditions, showing
 increasingly diverging trajectories (\autoref{fig:fig3}) for slightly different parameters, as well
 as unending oscillations (\autoref{fig:fig2}). The bifurcation diagrams (\autoref{fig:fig4}) further confirm
 the existence of chaos by illustrating the presence of cyclic behaviour for some values
-and chaotic intervals for others, hence the extreme sensibility of the system to $b_1$
+and chaotic intervals for others, hence the extreme sensitivity of the system to $b_1$
 values. As for the Poincaré sections (\autoref{fig:figS1} a,c), \citeauthor{hastings1991} plotted (x,y)
 coordinates of points of the phase plots that theoretically coincided with the plane in
 the "handle" of the teacup-shaped diagrams.
