+++ code/main.jl
@@ -15,7 +15,7 @@ using DifferentialEquations
 include("HPmodel.jl")
 
 # Make the figures
-for figfile in filter(f -> isfile(f)&startswith(f, "figure")&!endswith(f, "gif.jl"), readdir())
+for figfile in filter(f -> isfile(f) && startswith(f, "figure") && !endswith(f, "gif.jl"), readdir())
     @info "Running $figfile"
     include(figfile)
 end
