+++ code/figure5.jl
@@ -4,17 +4,17 @@
 #### Figure 5 (Poincaré section & map) ####
 #### b1 = 3.0
 
-u0=[0.76,0.16,9.9]  #initial conditions for x,y,z
-tspan=(0.0,10000.0) #timespan
-p_b3 = [5.0,0.1,3.0,2.0,0.4,0.01] # parameter values (a1, a2, b1, b2, c1, c2)
+u0 = [0.76, 0.16, 9.9]  #initial conditions for x,y,z
+tspan = (0.0, 10000.0) #timespan
+p_b3 = [5.0, 0.1, 3.0, 2.0, 0.4, 0.01] # parameter values (a1, a2, b1, b2, c1, c2)
 
 # Solving the system of differential equations
-prob_b3 = ODEProblem(parameterized_model,u0,tspan,p_b3)
-sol_b3 = solve(prob_b3, alg = RK4(), reltol=1e-14) # RK4 algorithme
+prob_b3 = ODEProblem(parameterized_model, u0, tspan, p_b3)
+sol_b3 = solve(prob_b3, alg = RK4(), reltol=1e-14)
 
-sol_x_b3 = sol_b3[1,:] # timeseries for x
-sol_y_b3 = sol_b3[2,:] # timeseries for y
-sol_z_b3 = sol_b3[3,:] # timeseries for z
+sol_x_b3 = sol_b3[1, :] # timeseries for x
+sol_y_b3 = sol_b3[2, :] # timeseries for y
+sol_z_b3 = sol_b3[3, :] # timeseries for z
 
 constant_b3 = 9.0 # poincaré section (z value)
 ϵ_b3 = 0.05 # margin of error for z value
@@ -35,8 +35,8 @@ for j in 1:length(ind_x_b3)
 end
 
 # for loop : non adjacent indices
-for j in 1:(length(ind_xyz_b3)-1)
-    if ind_xyz_b3[j] == ind_xyz_b3[j+1] - 1 # if 2 consecutive concordant indices are also
+for j in 1:(length(ind_xyz_b3) - 1)
+    if ind_xyz_b3[j] == ind_xyz_b3[j + 1] - 1 # if 2 consecutive concordant indices are also
                                       # consecutive in the solution, don't keep them
         ind_xyz_b3[j] = 0
     end
@@ -49,12 +49,12 @@ ind_xyz_b3 = ind_xyz_b3[findall(i -> i > 0, ind_xyz_b3)] # remove extra values (
 xmin = 0.95 ; xmax = .983 ; ymin = 0.015 ; ymax = .04
 fig5a = plot(sol_x_b3[ind_xyz_b3], sol_y_b3[ind_xyz_b3],
              legend = false,
-             seriestype=:scatter,
-             xlabel="x(n)", xlim=(xmin, xmax), xticks=xmin:0.01:xmax, xminorticks=2,
-             ylabel="x(n)", ylim=(ymin, ymax), yticks=ymin:0.01:ymax, yminorticks=2,
+             seriestype = :scatter,
+             xlabel = "x(n)", xlim = (xmin, xmax), xticks = xmin:0.01:xmax, xminorticks = 2,
+             ylabel = "x(n)", ylim = (ymin, ymax), yticks = ymin:0.01:ymax, yminorticks = 2,
              xgrid = :none, ygrid = :none,
-             title="A", titleloc=:left,
-             c=:black, ms=3)
+             title = "A", titleloc = :left,
+             c = :black, ms = 3)
 
 # Figure 5B - Poincaré map
 ind_n_b3 = ind_xyz_b3[1:length(ind_xyz_b3)-1] # indices for x(n)
@@ -63,27 +63,27 @@ ind_nplus1_b3 = ind_xyz_b3[2:length(ind_xyz_b3)] # indices for x(n+1)
[…]
