### 2020-08-12 78a0c68
add info message for package loading
