### 2020-08-12 f2b0b8b
specify model with more recent syntax
