### 2020-08-13 2c4b13b
fix linewidth on figure 2
