+++ article/content.tex
@@ -338,7 +338,7 @@ This reproduction was realized as part of the \textit{BIO6032 Biologie Computati
 
 \begin{figure}
   \centering
-  \includegraphics[scale=0.7]{figures/fig2.png} \caption{Time series of the nondimensional variables (a) $x$, (b) $y$ and (c) $z$, for $t$ ranging from 5000 to 6500 ($x$ = 0.76, $y$ = 0.16, and $z$ = 9.9 as initial conditions). The parameter values used in the simulations are given in \autoref{tbl:table1} ($b_1$ = 3.0). This figure replicates Fig. 2 A,B,C of \citet{hastings1991}.}
+  \includegraphics[scale=0.7]{figures/fig2.pdf} \caption{Time series of the nondimensional variables (a) $x$, (b) $y$ and (c) $z$, for $t$ ranging from 5000 to 6500 ($x$ = 0.76, $y$ = 0.16, and $z$ = 9.9 as initial conditions). The parameter values used in the simulations are given in \autoref{tbl:table1} ($b_1$ = 3.0). This figure replicates Fig. 2 A,B,C of \citet{hastings1991}.}
   \label{fig:fig2}
 \end{figure}
 
@@ -346,7 +346,7 @@ This reproduction was realized as part of the \textit{BIO6032 Biologie Computati
 
 \begin{figure}
   \centering
-  \includegraphics[scale=0.7]{figures/fig2d.png}
+  \includegraphics[scale=0.7]{figures/fig2d.pdf}
   \caption{Three-dimensional phase plot of species $x$, $y$ and $z$ for $t$ ranging from 1 to 10 000 ($x$ = 0.76, $y$ = 0.16, and $z$ = 9.9 as initial conditions). The parameter values used in the simulations are given in \autoref{tbl:table1} ($b_1$ = 3.0). This figure replicates Fig. 2D of \citet{hastings1991}.}
   \label{fig:fig2d}
 \end{figure}
@@ -355,7 +355,7 @@ This reproduction was realized as part of the \textit{BIO6032 Biologie Computati
 
 \begin{figure}
   \centering
-  \includegraphics[scale=0.7]{figures/fig3.png}
+  \includegraphics[scale=0.7]{figures/fig3.pdf}
   \caption{Time series of $x$, for $t$ ranging from 0 to 500. The solid and dashed lines have  $x$ = 0.77 and $x$ = 0.78 as initial conditions respectively ($y$ = 0.16 and $z$ = 9.9 as initial conditions are unchanged). The parameter values used in the simulations are given in \autoref{tbl:table1} ($b_1$ = 3.0). This figure replicates Fig. 3 of \citet{hastings1991}.}
   \label{fig:fig3}
 \end{figure}
@@ -364,7 +364,7 @@ This reproduction was realized as part of the \textit{BIO6032 Biologie Computati
 
 \begin{figure}
   \centering
-  \includegraphics[scale=0.7]{figures/fig4.png}
+  \includegraphics[scale=0.7]{figures/fig4.pdf}
   \caption{Bifurcation diagrams of the local maxima of $z$ plotted against $b_1$ ranging from (a) 2.2 to 3.2, (b) 3.0 to 6.2, and (c) 2.25 to 2.6. The other parameter values used in the simulations are given in \autoref{tbl:table1} ($x$ = 0.76, $y$ = 0.16, and $z$ = 9.9 as initial conditions). This figure replicates Fig. 4 of \citet{hastings1991}.}
   \label{fig:fig4}
 \end{figure}
@@ -373,7 +373,7 @@ This reproduction was realized as part of the \textit{BIO6032 Biologie Computati
 
 \begin{figure}
   \centering
[…]
