### 2020-08-14 523accc
add note on graphics display error in README
