+++ README.md
@@ -41,13 +41,19 @@ include("main.jl") # make sure you ran cd("code") to set the correct working dir
 
 Note that this is not an especially intensive task. On an ordinary laptop, it takes about 1 minute, 2.00 GB of memory and uses a single core. However, if using the packages for the first time, precompiling may take an additional 5 minutes (Julia has to do this after installing or changing versions).
 
-Alternatively, individual scripts in `code/` (producing one figure each) can be run separately after running lines 1-15 of `main.jl` (which will load the packages and the model). To run interactively, simply run the content line-by-line in a terminal or through a Julia IDE, such as [Juno](https://junolab.org/) or [Julia for VS Code](https://www.julia-vscode.org/).
-
-The GIF can be reproduced by running the following. This is longer than for the core figures and takes about 8-10 minutes.
-
+Alternatively, individual scripts in `code/` (producing one figure each) can be run separately, either line-by-line in a terminal or through a Julia IDE, such as [Juno](https://junolab.org/) or [Julia for VS Code](https://www.julia-vscode.org/). However, be sure to run these lines from `main.jl` first:
 
 ```julia
-include("figure2D-gif.jl") # make sure you ran cd("code") to set the correct working directory
+import Pkg; Pkg.activate(".")
+using DifferentialEquations
+using Plots
+include("HPmodel.jl")
+```
+
+The GIF can be reproduced by running the `figure2D-gif.jl` script. This is longer than for the core figures and takes about 8-10 minutes.
+
+```julia
+include("figure2D-gif.jl")
 ```
 
 #### Notes on possible warnings
