### 2020-08-14 21b32b4
add note on instable solution warning
