### 2020-08-14 8bd702b
remove mention of ParameterizedFunctions in text
