### 2020-08-14 ae53d7a
add note on integration algorithm choice
