### 2020-08-14 #15 @gabrieldansereau
Here are the changes I made to address @cboettig's comments (@rougier I guess you're acting as editor?). The full diff is [here](https://github.com/BIO6032/2019_replication_HastingsPowell_1991/compare/7b0e9930a1d4a4f307d119bade17eb8a895e5b55..ae53d7a289f7a3d7e8753c951baf1ba1f7a0a162) (I've also made some minor changes to make a few things clearer).

- [7b0e993](https://github.com/BIO6032/2019_replication_HastingsPowell_1991/commit/7b0e9930a1d4a4f307d119bade17eb8a895e5b55) - Clarifies how to run the scripts
- [6152ff5](https://github.com/BIO6032/2019_replication_HastingsPowell_1991/commit/6152ff579d2152df0a5cbb6825e1a529b0d0910c) - Clarifies the runtime & memory needs
- [e2ed990](https://github.com/BIO6032/2019_replication_HastingsPowell_1991/commit/e2ed99090be1b638da5e0674ab06a799d1bc9c60) - Fixes the sensitivity typo
- [6561093](https://github.com/BIO6032/2019_replication_HastingsPowell_1991/commit/6561093bd93a07dce187ca0053713ca56103d69d) - Replaces the figures by a sharper vectorized version
- [f2b0b8b ](https://github.com/BIO6032/2019_replication_HastingsPowell_1991/commit/f2b0b8b794c2a2dcbbb7b2cfcadbd2919a637d31)- Fixes the 3rd warning `Warning: Symbolic calculations could not initiate`
- [523accc](https://github.com/BIO6032/2019_replication_HastingsPowell_1991/commit/523accc0c8e85323fcc0e58336258dbcbf7629e5) - Adds a note on how to fix the 2nd warning `GKS: Open failed in routine OPEN_WS`. The figures should reproduce despite it. I believe it comes from external libraries and could be fixed on the user-side by [installing external dependencies](https://github.com/JuliaPlots/Plots.jl/issues/1649#issuecomment-592198276), but I cannot test it.
- [21b32b](https://github.com/BIO6032/2019_replication_HastingsPowell_1991/commit/21b32b42d1fe931a1ce17adab6509f1934f37170) - Adds a note to explain that the 1st warning `There is either an error in your model or the true solution is unstable` is expected. In the context of chaotic behaviour, we believed it made sense for the model not to converge and to be unstable for a few of the higher values tested for the b1 parameter. I see how it can be annoying as it's repeated many times, but I think it's better to leave it like that to warn potential users, rather than to find a way to hide it.

Regarding the choice of numerical integration algorithm, I've added a small note to the text in [ae53d7s](https://github.com/BIO6032/2019_replication_HastingsPowell_1991/commit/ae53d7a289f7a3d7e8753c951baf1ba1f7a0a162). Selecting a specific algorithm had an impact but did result in a similar behaviour for the system -- in fact, we had to use RK4 instead of the automatic solver to replicate some figures correctly (but it had its downsides too). Not knowing the original algorithm and initial conditions, we thought it best to start from the automatic algorithm if possible, then fiddle with the initials conditions to find the closest reproduction. So the initial conditions would not necessarily be the same with another algorithm.
