### 2020-08-17 #15 @rougier
@ChrisRackauckas @dmcglinn @jsta  @ha0ye Could any of you review [this submission (Ecology / Julia)](https://github.com/ReScience/submissions/issues/15)?
