### 2020-08-18 4fbc7ae
subscripts in figure 4 and 6 axis labels
