### 2020-08-18 5013f87
edits and comments to article
