### 2020-08-18 #15 @pboesu
@rougier @gabrieldansereau
I agree with @cboettig . This is an excellent concise replication. I only use julia very sporadically and had to install it from scratch on the machine used to review the code (Windows 8.1), but it was easy to run the code and reproduce the figures.

In addition to the warnings already discussed (and addressed by the authors) julia triggered a Windows Firewall warning about  "gksqt.exe", possibly an analogue to the ubuntu GKS warning.  Additionally there were compilation warnings for `DiffEqBase`:

```
WARNING: Method definition combine_axes(Any, Any) in module Broadcast at broadcast.jl:484 overwritten in module DiffEqBase at C:\Users\Philipp\.julia\packages\DiffEqBase\gTsoC\src\diffeqfastbc.jl:18.
  ** incremental compilation may be fatally broken for this module **
```
and when running the code in figure4.jl there were warnings about variable names:

```
┌ Warning: Assignment to `b1` in soft scope is ambiguous because a global variable by the same name exists: `b1` will be treated as a new local. Disambiguate by using `local b1` to suppress this warning or `global b1` to assign to the existing global variable.
└ @ C:\Users\Philipp\Documents\bitbucket\ReScience\2019_replication_HastingsPowell_1991\code\figure4.jl:22
┌ Warning: Assignment to `prob` in soft scope is ambiguous because a global variable by the same name exists: `prob` will be treated as a new local. Disambiguate by using `local prob` to suppress this warning or `global prob` to assign to the existing global variable.
└ @ C:\Users\Philipp\Documents\bitbucket\ReScience\2019_replication_HastingsPowell_1991\code\figure4.jl:24
┌ Warning: Assignment to `sol` in soft scope is ambiguous because a global variable by the same name exists: `sol` will be treated as a new local. Disambiguate by using `local sol` to suppress this warning or `global sol` to assign to the existing global variable.
```

Otherwise I only have a handfull of minor comments about the write-up and figures. I've created a pull request to the original repository that implements these suggestions: https://github.com/BIO6032/2019_replication_HastingsPowell_1991/pull/23

https://github.com/BIO6032/2019_replication_HastingsPowell_1991/commit/ef8143f47f5e920cfda81594cb6ce7b993cf13a6 fixes a typo in the axis label of Fig. 5

https://github.com/BIO6032/2019_replication_HastingsPowell_1991/commit/4fbc7ae8cffd8f45a5c6c9d1e9fbe7cc8975f6e6 is a suggestion to use LaTeXStrings for proper subscripts in Figure labels. This may require some minor refactoring to shift the new dependency to main.jl I didn't commit the changed *.toml files as I am unsure how to work with these. (Also this is a rather pedantic suggestion tbh - so I leave it up to the authors whether or not to incorporate this.)

https://github.com/BIO6032/2019_replication_HastingsPowell_1991/commit/5013f87b9382a82cd6a35aa815603c3e852f7295 includes comments and minor edits to the manuscript. A few comments that need additional revision by the authors are prefixed with a LaTeX comment string `%Reviewer comment:`

I believe this manuscript is in good shape to be published once these minor changes are implemented.
