### 2020-08-18 #15 @gabrieldansereau
Thank you very much for your review @pboesu.

I believe the soft scope warning came in more recent versions of Julia. Is it possible you're using 1.5 instead of 1.3.1?

Thanks for adding your comments in a pull-request. I am on vacation this week, so maybe one of my co-authors wants to handle it? @FrancisBanville @aammd @tpoisot
Otherwise I'll be happy to do it next week.
