+++ article/content.tex
@@ -99,13 +99,13 @@ As noted by \citeauthor{hastings1991}, numerical integration is the only way to 
 global dynamical behaviour of the system.
 We used \textit{Julia v1.3.1} \citep{bezanson2017}, along with packages \verb| DifferentialEquations.jl|
 \citep{rackauckas2017} to compute the numerical integrations and \verb|Plots.jl| to represent
-our results. 
-Not knowing the exact algorithm used in the original article, we let the \verb|solve()| function 
+our results.
+Not knowing the exact algorithm used in the original article, we let the \verb|solve()| function
 select the appropriate algorithm to solve our differential equations.
 In our implementation, it selected a composite algorithm combining, amongst others,
 algorithms Tsit5 and Rosenbrock23.
-We did note, however, that setting a specific algorithm such as RK4 sometimes resulted in a similar 
-behaviour for the system. 
+We did note, however, that setting a specific algorithm such as RK4 sometimes resulted in a similar
+behaviour for the system.
 
 To fully replicate key findings of the original paper, we focused on replicating
 original figures.
@@ -145,8 +145,8 @@ This raises the problem of correctly identifying the maximum values in the cycli
 dynamics. Moreover, \citeauthor{hastings1991} mention removing points resulting from
 secondary local maxima, but do not provide details on how to identify these points.
 Hence, we adopted the following method:
-1) we selected the 1000 last solutions for our system between time steps 1 and 10 000, in
-order to eliminate transient behaviour; %Reviewer comment: This is unclear to me. Do you mean time steps 9000-10000? Please clarify.
+1) we selected the solutions for the last 1000 evaluated time steps of our system, in
+order to eliminate transient behaviour;
 2) we selected the values that were greater than both their preceding and following
 values, which identified local maxima only;
 and 3) we only kept values that were greater than a given threshold of the cycle's maximal
@@ -235,8 +235,8 @@ Our additional bifurcation diagrams, where we varied $b_2$ instead of $b_1$ (\au
 confirm that chaos occurs for values other than $b_2$ = 2.0. Chaos is apparent for both
 smaller or greater values.
 However, while \citeauthor{hastings1991} reported that chaos was more likely for greater values of
-$b_2$, our results highlight that $z$ instead converges to a single value and starts to
-crash past $b_2$ = 2.35. %Reviewer comment: Please clarify what you mean by "crash" here. e.g. "starts to rapidly approach zero"
+$b_2$, our results highlight that $z$ instead converges to a single value and
+starts to rapidly approach zero past $b_2$ = 2.35.
 
 Lastly, although \citeauthor{hastings1991} did not specify the equation of the plane that
 crosses the trajectories of the phase plot at its "handle", we were able to accurately
@@ -293,8 +293,7 @@ at $b_1$ = 3.1.
 
 While we also found chaos for values of $b_2$ other than the default one of 2.0, both
 smaller or greater, we do not totally agree with \citeauthor{hastings1991} that "chaos is more
-likely for larger values of $b_2$". As \autoref{fig:figS1} shows %Reviewer comment: missing verb?
-, chaos can be quite likely for both
+likely for larger values of $b_2$". As \autoref{fig:figS1} shows, chaos can be quite likely for both
 smaller or larger values.
 We find important to note, however, that at a certain value of $b_2$, $z$ converges and
 starts to crash, thus exhibiting non chaotic behaviour within a given range of $b_1$
@@ -320,7 +319,7 @@ results confirm chaos arising in a three species food chain in continuous time.
 In general, the model, including its equations and parameters, was well described by the
 authors. The most significant obstacles to reproducibility in \citeauthor{hastings1991}'s paper
 were the absence of the values of the initial conditions, which have a huge impact on a
[…]
