### 2020-08-28 f15916f
fix ambiguous local variable warning
