### 2020-08-28 9fa0dd9
clarify explanation on 1000 solutions for fig 4
