+++ article/content.tex
@@ -145,8 +145,10 @@ This raises the problem of correctly identifying the maximum values in the cycli
 dynamics. Moreover, \citeauthor{hastings1991} mention removing points resulting from
 secondary local maxima, but do not provide details on how to identify these points.
 Hence, we adopted the following method:
-1) we selected the solutions for the last 1000 evaluated time steps of our system, in
-order to eliminate transient behaviour;
+1) we let the system run for time steps between 0 and 10 000, then kept the 1000 last 
+solutions to the numerical integration for $z$ to eliminate transient behaviour. 
+Note that these do not occur at time steps between 9000 and 10 000, as the system doesn't
+necessarily reach a stable solution, nor at the same time steps for all values of $b_1$;
 2) we selected the values that were greater than both their preceding and following
 values, which identified local maxima only;
 and 3) we only kept values that were greater than a given threshold of the cycle's maximal
