+++ article/content.tex
@@ -146,9 +146,9 @@ dynamics. Moreover, \citeauthor{hastings1991} mention removing points resulting 
 secondary local maxima, but do not provide details on how to identify these points.
 Hence, we adopted the following method:
 1) we let the system run for time steps between 0 and 10 000, then kept the 1000 last 
-solutions to the numerical integration for $z$ to eliminate transient behaviour. 
-Note that these do not occur at time steps between 9000 and 10 000, as the system doesn't
-necessarily reach a stable solution, nor at the same time steps for all values of $b_1$;
+solutions to the numerical integration for $z$ to eliminate transient behaviour 
+(note that these do not occur at time steps between 9000 and 10 000, as the system doesn't
+necessarily reach a stable solution, and that the exact time steps vary for all values of $b_1$);
 2) we selected the values that were greater than both their preceding and following
 values, which identified local maxima only;
 and 3) we only kept values that were greater than a given threshold of the cycle's maximal
