### 2020-08-31 d1dd692
add minor detail on continuous-time
