+++ article/content.tex
@@ -36,7 +36,7 @@ All code used to replicate the original model is available alongside this articl
 \section{Methods}
 
 The model formulation used in this paper is the same as in the original
-publication. \citeauthor{hastings1991} used a 14 parameter model to represent the three-species
+publication. \citeauthor{hastings1991} used a continuous-time model with 14 parameters to represent the three-species
 food chain, with $X$, $Y$, and $Z$ as the numbers of the species at the lowest level of
 the food chain, of the species that preys upon $X$, and of the species that preys upon
 $Y$, respectively.
@@ -145,11 +145,11 @@ This raises the problem of correctly identifying the maximum values in the cycli
 dynamics. Moreover, \citeauthor{hastings1991} mention removing points resulting from
 secondary local maxima, but do not provide details on how to identify these points.
 Hence, we adopted the following method:
-1) we let the system run for time steps between 0 and 10 000, then kept the 1000 last 
-solutions to the numerical integration for $z$ to eliminate transient behaviour 
+1) we let the system run for continuous time steps between 0 and 10 000, then kept the last 1000 
+solutions to the numerical integration to eliminate transient behaviour 
 (note that these do not occur at time steps between 9000 and 10 000, as the system doesn't
 necessarily reach a stable solution, and that the exact time steps vary for all values of $b_1$);
-2) we selected the values that were greater than both their preceding and following
+2) we selected the values of $z$ that were greater than both their preceding and following
 values, which identified local maxima only;
 and 3) we only kept values that were greater than a given threshold of the cycle's maximal
 amplitude, in order to remove secondary local maxima.
