### 2020-08-31 d693629
fix a few typos
