### 2020-08-31 #15 @gabrieldansereau
@rougier Sorry for the delay.

Thanks again for your review @pboesu

We merged your pull-request and added these changes:
- BIO6032/2019_replication_HastingsPowell_1991@c9c8603 - Addresses the comments you left with `%Reviewer comment:`. The one on L149 regarding the time steps used for figure 4 was a bit more complex and changed a few times, last in BIO6032/2019_replication_HastingsPowell_1991@d1dd692
- BIO6032/2019_replication_HastingsPowell_1991@f15916f2 - Fixes the warning about variables in soft scope with Julia 1.5
- BIO6032/2019_replication_HastingsPowell_1991@3c1f463 - Adds `LatexStrings` to the *.toml files

I think the "gksqt.exe" warning is indeed similar to the Ubuntu GKS one. I'd leave it as is since it's user-specific and does not stop reproduction. Same for the `DiffEqBase` precompilation warning, which does not seem to impact either.
