### 2020-09-01 #15 @rougier
Nice. @cboettig @pboesu Are tyou satisfied with the corrections and answers provided by @gabrieldansereau. If yes (you can add a thumb up), we can proceed with publication.
