### 2020-09-01 #15 @cboettig
Yes

On Tue, Sep 1, 2020 at 12:33 AM Nicolas P. Rougier <notifications@github.com>
wrote:

>
>
> Nice. @cboettig <https://github.com/cboettig> @pboesu
> <https://github.com/pboesu> Are tyou satisfied with the corrections and
> answers provided by @gabrieldansereau
> <https://github.com/gabrieldansereau>. If yes (you can add a thumb up),
> we can proceed with publication.
>
>
>
>
