+++ article/metadata.tex
@@ -1,15 +1,15 @@
 % DO NOT EDIT - automatically generated from metadata.yaml
 
 \def \codeURL{https://github.com/BIO6032/2019_replication_HastingsPowell_1991/}
-\def \codeDOI{}
+\def \codeDOI{10.5281/zenodo.4012773}
 \def \dataURL{}
 \def \dataDOI{}
-\def \editorNAME{}
-\def \editorORCID{}
-\def \reviewerINAME{}
-\def \reviewerIORCID{}
-\def \reviewerIINAME{}
-\def \reviewerIIORCID{}
+\def \editorNAME{Nicolas P. Rougier}
+\def \editorORCID{0000-0002-6972-589X}
+\def \reviewerINAME{Carl Boettiger}
+\def \reviewerIORCID{0000-0002-1642-628X}
+\def \reviewerIINAME{Philipp Boersch-Supan}
+\def \reviewerIIORCID{0000-0001-6723-6833}
 \def \dateRECEIVED{}
 \def \dateACCEPTED{}
 \def \datePUBLISHED{}
@@ -18,14 +18,14 @@
 \def \articleDOMAIN{ecology}
 \def \articleBIBLIOGRAPHY{bibliography.bib}
 \def \articleYEAR{2020}
-\def \reviewURL{}
+\def \reviewURL{https://github.com/ReScience/submissions/issues/15}
 \def \articleABSTRACT{ Hastings \& Powell's model describing chaos in a three-species food chain is a classic in ecological modelling. Here, we provide a successul reproduction of the model and results from the original article, implemented in Julia v1.3.1 and openly accessible online. The main challenge we faced was the absence of the initial conditions from the original article. Our reproduction nonetheless displays the same results, although with slight differences, and confirms chaos arising in a three species food chain in continuous time. }
 \def \replicationCITE{Hastings, A., & Powell, T. (1991). Chaos in a Three-Species Food Chain. Ecology, 72(3), 896–903. https://doi.org/10.2307/1940591}
 \def \replicationBIB{hastings1991}
 \def \replicationURL{https://esajournals.onlinelibrary.wiley.com/doi/abs/10.2307/1940591}
 \def \replicationDOI{10.2307/1940591}
-\def \contactNAME{Timothée Poisot}
-\def \contactEMAIL{timothee.poisot@umontreal.ca}
+\def \contactNAME{Gabriel Dansereau}
+\def \contactEMAIL{gabriel.dansereau@umontreal.ca}
 \def \articleKEYWORDS{rescience, ecological modelling, food chain, chaos, julia}
 \def \journalNAME{None}
 \def \journalVOLUME{}
