+++ article/content.tex
@@ -336,6 +336,8 @@ obtain an even better performing implementation.
 GD, FB and EB performed the first version of the reproduction and wrote the first version of the manuscript. AMD and TP provided guidance and comments during the reproduction, and revised the manuscript. GD performed the final revisions.
 
 \section{Acknowledgments}
+We thank reviewers Carl Boettiger and Philipp Boersch-Supan for their helpful comments to ensure the
+reproducibility of this project.
 This reproduction was realized as part of the \textit{BIO6032 Biologie Computationnelle et Modélisation} course at Université de Montréal.
 
 \newpage
