+++ article/metadata.tex
@@ -1,7 +1,7 @@
 % DO NOT EDIT - automatically generated from metadata.yaml
 
 \def \codeURL{https://github.com/BIO6032/2019_replication_HastingsPowell_1991/}
-\def \codeDOI{10.5281/zenodo.4012773}
+\def \codeDOI{10.5281/zenodo.4012772}
 \def \dataURL{}
 \def \dataDOI{}
 \def \editorNAME{Nicolas P. Rougier}
