+++ article/metadata.yaml
@@ -51,7 +51,7 @@ keywords: rescience, ecological modelling, food chain, chaos, julia
 #   see https://guides.github.com/activities/citable-code/
 code:
   - url: https://github.com/BIO6032/2019_replication_HastingsPowell_1991/
-  - doi: 10.5281/zenodo.4012773
+  - doi: 10.5281/zenodo.4012772
 
 # Date URL and DOI (optional if no data)
 data:
