### 2020-09-03 #15 @gabrieldansereau
@rougier Hope it's ok -- I went ahead and updated the metadata.yaml file. I added the DOI for the code, the review URL, and the reviewers' info BIO6032/2019_replication_HastingsPowell_1991@3a46622
