### 2020-09-08 #15 @rougier
@gabrieldansereau That's perfect. Let check for the volume/issue and I will send you the DOI (or make a PR on your repo)
