### 2020-09-10 #15 @rougier
I've upload an edited version at https://sandbox.zenodo.org/record/668390. This is the Zenodo **sandbox** (entry will dissappear in one day). Can you check everything's fine ? If yes, I'll publish on the actual server
