### 2020-09-10 87b3051
Edited metadata for pulication + 1 correction in latex math env
