### 2020-09-10 #15 @rougier
It's published at https://zenodo.org/record/4022518 and will appear on the ReScience website in a few minutes. Thanks again for your submission and congratulation again ! I've made a PR with publication data on your repo.

By the way, I forgot to ask because you were too fast in updating the metadata but now the preferred and recommended way to save your code is to use software heritage  (https://www.softwareheritage.org/save-and-reference-research-software/).
