### 2020-09-10 #15 @gabrieldansereau
Great, many thanks to you! I've merged your PR on our repo.

I did not know about Software Heritage, I'll check it out!
